third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-block-import.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="script-src 'unsafe-inline'">
<link id="target" rel="import" href="resources/hello.html">
</head>
<body>
<pre id="result"></pre>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

if (window.target.import == null)
    result.innerHTML = "PASS\n";
else
    result.innerHTML = "FAIL: The import should be blocked.\n";
</script>
</body>
</html>
