third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-block-import-expected.txt

CONSOLE ERROR: line 5: Refused to load the script 'http://127.0.0.1:8000/htmlimports/resources/hello.html' because it violates the following Content Security Policy directive: "script-src 'unsafe-inline'".

PASS
